% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca.R
\name{busca}
\alias{busca}
\title{Scrapea las entradas resultantes de una búsqueda en meneame.com}
\usage{
busca(palabra, paginas, ruta = "~/extraccion.csv")
}
\arguments{
\item{paginas}{Numeric. Número de páginas que se quieren scraper de los resultados de la búsqueda. Como máximo 40.}

\item{ruta}{Character. Ruta en el ordenador donde se quiere el csv.}

\item{sub}{Character. Palabra que se quiere buscar en meneame.com que se quiere scrapear.}
}
\value{
Un csv que se guarda en la ruta indicada.
}
\description{
Scrapea las entradas resultantes de una búsqueda en meneame.com
}
