% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submeneame.R
\name{submeneame}
\alias{submeneame}
\title{Scrapea las entradas en un sub de meneame.com}
\usage{
submeneame(sub, paginas, ruta = "~/extraccion.csv")
}
\arguments{
\item{sub}{Character. Sub de meneame.com que se quiere scrapear.}

\item{paginas}{Numeric. Número de páginas que se quieren scraper. No pone más de la que haya.}

\item{ruta}{}
}
\value{
A csv
}
\description{
Scrapea las entradas en un sub de meneame.com
}
