% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portada.R
\name{portada}
\alias{portada}
\title{Scrapea las entradas en la portada de meneame.com}
\usage{
portada(paginas, ruta = "~/extraccion.csv")
}
\arguments{
\item{paginas}{Numeric. Número de páginas que se quieren scraper. No pone más de la que haya.}

\item{ruta}{}
}
\value{
A csv
}
\description{
Scrapea las entradas en la portada de meneame.com
}
